feat: Update handleSelectPlace to update user places on server
The code changes in this commit update the `handleSelectPlace` function in the `App` component to also update the user places on the server. This is done by making an asynchronous call to the `updateUserPlaces` function from the `http.js` module. If there is an error during the update, the user places are reverted to their previous state and an error message is logged to the console.